+++ packages/mst-middlewares/package.json
@@ -1,37 +1,27 @@
 {
-  "name": "middleware",
+  "name": "mst-middlewares",
   "version": "1.0.0",
   "description": "The MST package ships with some prebuilt middlewares, which serves mainly as examples on how to write your own middleware. The source of each middleware can be found in this github directory, you are encouraged to read them!",
   "main": "atomic.js",
   "scripts": {
-    "test": "ava",
-    "start": "ava -w"
+    "test": "tsc -p test/ && ava",
+    "build": "tsc"
   },
   "author": "",
-  "license": "ISC",
+  "license": "MIT",
   "devDependencies": {
     "ava": "^0.22.0",
-    "babel-plugin-transform-regenerator": "^6.26.0",
-    "babel-polyfill": "^6.26.0"
+    "mobx-state-tree": "^1.0.1"
+  },
+  "peerDependencies": {
+    "mobx-state-tree": "^1.0.1"
   },
   "ava": {
     "files": [
-      "*.spec.js"
+      "test-lib/test/*.js"
     ],
     "source": [
-      "*.js"
-    ],
-    "babel": "inherit",
-    "require": [
-      "babel-register"
-    ]
-  },
-  "babel": {
-    "presets": [
-      "es2015"
-    ],
-    "plugins": [
-      "transform-regenerator"
+      "test-lib/src/**/*.js"
     ]
   }
 }
