+++ packages/mst-middlewares/src/simple-action-logger.ts
@@ -1,5 +1,3 @@
-"use strict"
-
 import { getPath } from "mobx-state-tree"
 
 export default function simpleActionLogger(call, next) {
