+++ packages/mst-middlewares/test/tsconfig.json
@@ -1,0 +1,16 @@
+{
+    "compilerOptions": {
+        "rootDir": "../",
+        "module": "commonjs",
+        "target": "es2015",
+        "noImplicitAny": false,
+        "noImplicitThis": true,
+        "experimentalDecorators": true,
+        "strictNullChecks": true,
+        "sourceMap": false,
+        "outDir": "../test-lib"
+    },
+    "include": [
+        "**/*.ts"
+    ]
+}
