+++ packages/mst-middlewares/tsconfig.json
@@ -1,0 +1,25 @@
+{
+    "version": "2.0.0",
+    "compilerOptions": {
+        "target": "es5",
+        "sourceMap": true,
+        "declaration": true,
+        "module": "es2015",
+        "removeComments": false,
+        "outDir": "lib/",
+        "noImplicitAny": true,
+        "moduleResolution": "node",
+        "experimentalDecorators": true,
+        "strictNullChecks": true,
+        "noFallthroughCasesInSwitch": true,
+        "noImplicitReturns": true,
+        "noImplicitThis": true,
+        "importHelpers": true,
+        "lib": [
+            "dom", "es2015", "scripthost", "es2015.promise", "es2015.generator", "es2015.iterable"
+        ]
+    },
+    "files": [
+        "src/index.ts"
+    ]
+}
